requirements.txt
requests==2.31.0
wafw00f==2.2.0
urllib3==2.2.1
notify-py==0.3.42
pync==2.0.3
bs4
httpx
pyyaml